requirements.txt
streamlit==0.67.0
pandas==1.1.2
Pillow==7.2.0
scikit_learn==0.23.2
gunicorn==20.0.4
numpy>=1.18.5
scipy>=1.5.2






